+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 0.2.0 - XX XXXX 2020
 --------------------
 
+- Lookup listeners are just simple callables now.
 - Follows PEP 561 for packages providing typing information.
 
 0.1.0 - 18 May 2019